requirement.txt
bintrees==2.1.0
certifi>=2020.4.5
chardet>=3.0.4
click>=7.1.2
dateutils>=0.6.8
deap>=1.4.1
Flask>=1.1.2
gevent>=20.6.1
greenlet>=0.4.16
idna>=2.9
itsdangerous>=1.1.0
Jinja2>=2.11.2
MarkupSafe>=1.1.1
numpy>=1.18.5
pyaml>=20.4.0
python-dateutil>=2.8.1
pytz>=2020.1
PyYAML>=5.3.1
requests>=2.23.0
six>=1.15.0
sortedcontainers>=2.2.2
SQLAlchemy>=1.4.0,<2.1.0
tulipy>=0.4.0
redis>=5.0.0
influxdb-client>=1.38.0
hiredis>=2.2.0
urllib3>=1.25.9
websocket>=0.2.1
websocket-client>=0.57.0
Werkzeug>=1.0.1
zope.event>=4.4
zope.interface>=5.1.0
